fix(trading-engine): bound OrderBook's three collections, add endurance test
Closes the root cause of the 2026-07 GC death spiral. All three collections
in OrderBook grew without limit — three writes in addOrder(), zero removals —
which accumulated 11.36M Order objects over 7.6 days, exhausted the heap and
left the JVM in back-to-back Full GCs that could reclaim nothing.

Retained memory is now O(retention) instead of O(submissions).

## What changed

allOrders -> recentOrders, a rolling window, plus AtomicLong counters.
The RCA proposed replacing the list with a counter outright; that turned out
to be wrong. TradingApiServer.java:108 pages over getAllOrders() and
TradingFlowIntegrationTest filters it by BUY/SELL, so removing it breaks both.
Keeping a window alongside exact all-time counters satisfies both needs:
totalOrderCount() and uniqueOrderCount() stay exact through eviction because
they no longer depend on retained objects.

orders and orderIdFrequency are evicted together, oldest id first, so the
frequency map can never outgrow the id map.

Both counters return long. Overflow at 17.29 orders/sec would take 3.9 years
on int, which is precisely the timescale this class of defect lives on.

## Why not LinkedHashMap/removeEldestEntry

That is the pattern OrderCache uses and the RCA suggested, but it is not
thread-safe, and wrapping it in synchronizedMap would expose
getOrderIdFrequency() — serialized directly at TradingApiServer.java:182 — to
unsynchronized iteration. That is exactly BUG-11 again. ConcurrentHashMap
iterators are weakly consistent, so the maps stay concurrent and a small lock
covers only the two deques.

## Why 10,000 is enough

generateOrder() creates duplicates via stepsBack = (random.nextInt(3) + 1) * 2,
so a duplicate id is never more than 6 submissions behind its original. The
default retention leaves roughly 1600x margin, meaning eviction cannot degrade
duplicate detection at the engine's actual behaviour. An id evicted and later
resubmitted would read as new; at distance <= 6 that cannot happen.

## Endurance test

OrderBookRetentionTest asserts the property functional tests structurally
cannot: when input volume grows 10x, does the retained set grow 10x?

Verified it actually catches the defect. With retention set to
Integer.MAX_VALUE to reproduce unbounded behaviour, retained entries went
1,000 -> 100,000 and post-GC heap went 29 MB -> 70 MB; both assertions failed
as they should.

The first attempt at that control used `while (false)` to disable eviction.
That is an unreachable-statement compile error, so the build failed and the
surefire report being read was stale from the previous run — it looked like a
pass. Worth recording: when proving a test can fail, confirm the build
succeeded first.

Heap measurement collects repeatedly and verifies via GarbageCollectorMXBean
that collection actually occurred, failing rather than reporting a false pass
if the JVM ignored System.gc().

## Verification

- 66 trading-engine tests (63 + 3 endurance) and 43 payment-api tests pass.
- The one failure, DBValidationTest.buySellRatioIsBalanced, is pre-existing
  and unrelated: it queries the live MySQL table still holding the incident's
  11,358,422 rows. Skipped in CI via assumeTrue.
- tools/check-bounded-collections.sh violations: 8 -> 5. The remaining five
  are the payment-api and websocket fields, out of scope here.

Docs updated so the status claims stay honest: the RCA's "remediation not
implemented" note now records what landed, what deviated from the spec and
why, and what is still outstanding.

+++ trading-engine-simulator/src/test/java/com/binance/trading/integration/TradingFlowIntegrationTest.java
@@ -82,8 +82,8 @@ class TradingFlowIntegrationTest {
         assertFalse(duplicates.isEmpty(),
             "Pattern 1 FAIL: 20% duplicate rate → expect duplicates after 2s, got 0");
 
-        int uniqueCount = engine.getOrderBook().uniqueOrderCount();
-        int totalCount  = engine.getOrderBook().totalOrderCount();
+        long uniqueCount = engine.getOrderBook().uniqueOrderCount();
+        long totalCount  = engine.getOrderBook().totalOrderCount();
         assertTrue(totalCount > uniqueCount,
             "Pattern 1 FAIL: total(" + totalCount + ") must exceed unique(" + uniqueCount + ")");
